+++ cmd/autonode/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var version = "0.3.1"
+var version = "0.4.0"
 
 func main() {
 	// Create root command